Finsh Golang Docker for Prod and Dev
Using `dep` we add vendored sources, this allows us to build images
without any extra steps. We then install binaries to /usr/local/bin
and launch them via supervisord.

This means we don't have any source code on the box, allowing us to
use the same images in prod/dev. If you do want source code on the
box use docker-sync

The only requirement for a developer is that `dep` must be installed
on the host and the vendor must be inited. Whole process will be
documented in the wiki and refined